+++ src/cz/vutbr/fit/tam/paint3d/PaintingPoint.java
@@ -5,11 +5,11 @@ class PaintingPoint {
     public static final String TAG = "3Dpaint|PaintingPoint";
     public static final String TABLE_NAME = "painting_point";
     public static final String[] TABLE_COLUMNS = {"painting_point_id", "painting_id", "x", "y", "z"};
-    public int x;
-    public int y;
-    public int z;
+    public float x;
+    public float y;
+    public float z;
 
-    public PaintingPoint(int x, int y, int z) {
+    public PaintingPoint(float x, float y, float z) {
         this.x = x;
         this.y = y;
         this.z = z;
